backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
python-multipart
sqlalchemy
pydantic
passlib
python-jose